+++ apis/securesourcemanager/v1alpha1/types.generated.go
@@ -14,10 +14,6 @@
 
 package v1alpha1
 
-import (
-	refs "github.com/GoogleCloudPlatform/k8s-config-connector/apis/refs/v1beta1"
-)
-
 // +kcc:proto=google.cloud.securesourcemanager.v1.Instance.HostConfig
 type Instance_HostConfig struct {
 	// Output only. HTML hostname.
@@ -32,24 +28,6 @@ type Instance_HostConfig struct {
 
 	// Output only. Git SSH hostname.
 	GitSSH *string `json:"gitSSH,omitempty"`
-}
-
-// +kcc:proto=google.cloud.securesourcemanager.v1.Instance.PrivateConfig
-type Instance_PrivateConfig struct {
-	// Required. Immutable. Indicate if it's private instance.
-	IsPrivate *bool `json:"isPrivate,omitempty"`
-
-	// Required. Immutable. CA pool resource, resource must in the format of
-	//  `projects/{project}/locations/{location}/caPools/{ca_pool}`.
-	CaPoolRef *refs.PrivateCACAPoolRef `json:"caPoolRef,omitempty"`
-
-	// Output only. Service Attachment for HTTP, resource is in the format of
-	//  `projects/{project}/regions/{region}/serviceAttachments/{service_attachment}`.
-	HTTPServiceAttachment *string `json:"httpServiceAttachment,omitempty"`
-
-	// Output only. Service Attachment for SSH, resource is in the format of
-	//  `projects/{project}/regions/{region}/serviceAttachments/{service_attachment}`.
-	SSHServiceAttachment *string `json:"sshServiceAttachment,omitempty"`
 }
 
 // +kcc:proto=google.cloud.securesourcemanager.v1.Repository
